# [bugfix] Fix startOf/endOf DST issues while boosting performance