--- a/src/lib/create/date-from-array.js
+++ b/src/lib/create/date-from-array.js
@@ -1,12 +1,16 @@
 export function createDate (y, m, d, h, M, s, ms) {
-    // the date constructor remaps years 0-99 to 1900-1999
-    var remapYears = (y < 100 && y >= 0);
     // can't just apply() to create a date:
     // https://stackoverflow.com/q/181348
-    var date = new Date(remapYears ? y + 400 : y, m, d, h, M, s, ms);
-
-    if (remapYears && isFinite(date.getFullYear())) {
-        date.setFullYear(y);
+    var date;
+    // the date constructor remaps years 0-99 to 1900-1999
+    if (y < 100 && y >= 0) {
+        // preserve leap years using a full 400 year cycle, then reset
+        date = new Date(y + 400, m, d, h, M, s, ms);
+        if (isFinite(date.getFullYear())) {
+            date.setFullYear(y);
+        }
+    } else {
+        date = new Date(y, m, d, h, M, s, ms);
     }
 
     return date;
@@ -17,6 +21,7 @@ export function createUTCDate (y) {
     // the Date.UTC function remaps years 0-99 to 1900-1999
     if (y < 100 && y >= 0) {
         var args = Array.prototype.slice.call(arguments);
+        // preserve leap years using a full 400 year cycle, then reset
         args[0] = y + 400;
         date = new Date(Date.UTC.apply(null, args));
         if (isFinite(date.getUTCFullYear())) {