--- a/src/lib/moment/start-end-of.js
+++ b/src/lib/moment/start-end-of.js
@@ -5,10 +5,29 @@ var MS_PER_SECOND = 1000;
 var MS_PER_MINUTE = 60 * MS_PER_SECOND;
 var MS_PER_HOUR = 60 * MS_PER_MINUTE;
 var MS_PER_24_HOUR_DAY = 24 * MS_PER_HOUR;
+var MS_PER_400_YEARS = (365 * 400 + 97) * MS_PER_24_HOUR_DAY;
 
 // actual modulo - handles negative numbers (for dates before 1970):
 function mod(dividend, divisor) {
     return (dividend % divisor + divisor) % divisor;
+}
+
+function localStartOfDate(y, m, d) {
+    // the date constructor remaps years 0-99 to 1900-1999
+    if (0 <= y && y <= 99) {
+        return new Date(y + 400, m, d) - MS_PER_400_YEARS;
+    } else {
+        return new Date(y, m, d).getTime();
+    }
+}
+
+function utcStartOfDate(y, m, d) {
+    // Date.UTC remaps years 0-99 to 1900-1999
+    if (0 <= y && y <= 99) {
+        return Date.UTC(y + 400, m, d) - MS_PER_400_YEARS;
+    } else {
+        return Date.UTC(y, m, d);
+    }
 }
 
 export function startOf (units) {
@@ -18,56 +37,32 @@ export function startOf (units) {
         return this;
     }
 
+    var startOfDate = this._isUTC ? utcStartOfDate : localStartOfDate;
+
     switch (units) {
         case 'year':
-            time = (
-                (this._isUTC ?
-                    Date.UTC(this.year(), 0) :
-                    new Date(this.year(), 0).valueOf())
-            );
+            time = startOfDate(this.year(), 0, 1);
             break;
         case 'quarter':
-            time = (
-                (this._isUTC ?
-                    Date.UTC(this.year(), this.month() - this.month() % 3) :
-                    new Date(this.year(), this.month() - this.month() % 3).valueOf())
-            );
+            time = startOfDate(this.year(), this.month() - this.month() % 3, 1);
             break;
         case 'month':
-            time = (
-                (this._isUTC ?
-                    Date.UTC(this.year(), this.month()) :
-                    new Date(this.year(), this.month()).valueOf())
-            );
+            time = startOfDate(this.year(), this.month(), 1);
             break;
         case 'week':
-            // start of week should be: this.date() - this.weekday():
-            time = (
-                (this._isUTC ?
-                    Date.UTC(this.year(), this.month(), this.date() - this.weekday()) :
-                    new Date(this.year(), this.month(), this.date() - this.weekday()).valueOf())
-            );
+            time = startOfDate(this.year(), this.month(), this.date() - this.weekday());
             break;
         case 'isoWeek':
-            // start of iso week should be: this.date() - (this.isoWeekday - 1):
-            time = (
-                (this._isUTC ?
-                    Date.UTC(this.year(), this.month(), this.date() - (this.isoWeekday() - 1)) :
-                    new Date(this.year(), this.month(), this.date() - (this.isoWeekday() - 1)).valueOf())
-            );
+            time = startOfDate(this.year(), this.month(), this.date() - (this.isoWeekday() - 1));
             break;
         case 'day':
         case 'date':
-            time = this._d.valueOf();
-            time = (
-                (this._isUTC ?
-                    time - mod(time, MS_PER_24_HOUR_DAY) :
-                    new Date(this.year(), this.month(), this.date()).valueOf())
-            );
+            time = startOfDate(this.year(), this.month(), this.date());
             break;
         case 'hour':
             time = this._d.valueOf();
             time = (time - mod(time + (this._isUTC ? 0 : this.utcOffset() * MS_PER_MINUTE), MS_PER_HOUR));
+            // FIXME: Special case: handle fractional DST changes:
             break;
         case 'minute':
             time = this._d.valueOf();
@@ -91,56 +86,32 @@ export function endOf (units) {
         return this;
     }
 
+    var startOfDate = this._isUTC ? utcStartOfDate : localStartOfDate;
+
     switch (units) {
         case 'year':
-            time = (
-                (this._isUTC ?
-                    Date.UTC(this.year() + 1, 0) :
-                    new Date(this.year() + 1, 0).valueOf()) - 1
-            );
+            time = startOfDate(this.year() + 1, 0, 1) - 1;
             break;
         case 'quarter':
-            time = (
-                (this._isUTC ?
-                    Date.UTC(this.year(), 3 + this.month() - this.month() % 3) :
-                    new Date(this.year(), 3 + this.month() - this.month() % 3).valueOf()) - 1
-            );
+            time = startOfDate(this.year(), this.month() - this.month() % 3 + 3, 1) - 1;
             break;
         case 'month':
-            time = (
-                (this._isUTC ?
-                    Date.UTC(this.year(), 1 + this.month()) :
-                    new Date(this.year(), 1 + this.month()).valueOf()) - 1
-            );
+            time = startOfDate(this.year(), this.month() + 1, 1) - 1;
             break;
         case 'week':
-            // start of week should be: this.date() - this.weekday(), so end of week would be: this.date() - this.weekday() + 7
-            time = (
-                (this._isUTC ?
-                    Date.UTC(this.year(), this.month(), this.date() + 7 - this.weekday()) :
-                    new Date(this.year(), this.month(), this.date() + 7 - this.weekday()).valueOf()) - 1
-            );
+            time = startOfDate(this.year(), this.month(), this.date() - this.weekday() + 7) - 1;
             break;
         case 'isoWeek':
-            // start of iso week should be: this.date() - (this.isoWeekday - 1), so end would be: this.date() - (this.isoWeekday() - 1) + 7
-            time = (
-                (this._isUTC ?
-                    Date.UTC(this.year(), this.month(), this.date() + 7 - (this.isoWeekday() - 1)) :
-                    new Date(this.year(), this.month(), this.date() + 7 - (this.isoWeekday() - 1)).valueOf()) - 1
-            );
+            time = startOfDate(this.year(), this.month(), this.date() - (this.isoWeekday() - 1) + 7) - 1;
             break;
         case 'day':
         case 'date':
-            time = this._d.valueOf();
-            time = (
-                (this._isUTC ?
-                    time - mod(time, MS_PER_24_HOUR_DAY) + MS_PER_24_HOUR_DAY :
-                    new Date(this.year(), this.month(), this.date() + 1).valueOf()) - 1
-            );
+            time = startOfDate(this.year(), this.month(), this.date() + 1) - 1;
             break;
         case 'hour':
             time = this._d.valueOf();
             time = (time - mod(time + (this._isUTC ? 0 : this.utcOffset() * MS_PER_MINUTE), MS_PER_HOUR) + MS_PER_HOUR - 1);
+            // FIXME: Special case: handle fractional DST changes:
             break;
         case 'minute':
             time = this._d.valueOf();