--- a/src/lib/moment/start-end-of.js
+++ b/src/lib/moment/start-end-of.js
@@ -14,91 +14,168 @@ function mod(dividend, divisor) {
 export function startOf (units) {
     var time;
     units = normalizeUnits(units);
-    // the following switch intentionally omits break keywords
-    // to utilize falling through the cases.
-    switch (units) {
-        case 'year':
-            this.month(0);
-            /* falls through */
-        case 'quarter':
-        case 'month':
-            this.date(1);
-            /* falls through */
-        case 'week':
-        case 'isoWeek':
-        case 'day':
-        case 'date':
-            if (this.isValid()) {
-                if (this._isUTC) {
-                    time = this._d.valueOf();
-                    this._d.setTime(time - mod(time, MS_PER_24_HOUR_DAY));
-                } else {
-                    this._d.setTime(new Date(this.year(), this.month(), this.date()).valueOf());
-                }
+    if (units === undefined || units === 'millisecond' || !this.isValid()) {
+        return this;
+    }
+
+    if (this._isUTC) {
+        switch (units) {
+            case 'year':
+                this._d.setTime(Date.UTC(this.year(), 0));
                 hooks.updateOffset(this, true);
-            }
-            break;
-        case 'hour':
-            if (this.isValid()) {
-                var offset = this._isUTC ? 0 : this.utcOffset();
+                break;
+            case 'quarter':
+                this._d.setTime(Date.UTC(this.year(), this.month() - (this.month() % 3)));
+                hooks.updateOffset(this, true);
+                break;
+            case 'month':
+                this._d.setTime(Date.UTC(this.year(), this.month()));
+                hooks.updateOffset(this, true);
+                break;
+            case 'week':
+            case 'isoWeek':
+            case 'day':
+            case 'date':
                 time = this._d.valueOf();
-                this._d.setTime(time - mod(time + offset * MS_PER_MINUTE, MS_PER_HOUR));
+                this._d.setTime(time - mod(time, MS_PER_24_HOUR_DAY));
                 hooks.updateOffset(this, true);
-            }
-            break;
-        case 'minute':
-            if (this.isValid()) {
+                // Note: there may be a better/faster way of doing this with UTC...
+                if (units === 'week') {
+                    this.weekday(0);
+                } else if (units === 'isoWeek') {
+                    this.isoWeekday(1);
+                }
+                break;
+            case 'hour':
+                time = this._d.valueOf();
+                this._d.setTime(time - mod(time, MS_PER_HOUR));
+                hooks.updateOffset(this, true);
+                break;
+            case 'minute':
                 time = this._d.valueOf();
                 this._d.setTime(time - mod(time, MS_PER_MINUTE));
-            }
-            break;
-        case 'second':
-            if (this.isValid()) {
+                break;
+            case 'second':
                 time = this._d.valueOf();
                 this._d.setTime(time - mod(time, MS_PER_SECOND));
-            }
-            break;
-    }
-
-    // weeks are a special case
-    if (units === 'week') {
-        this.weekday(0);
-    }
-    if (units === 'isoWeek') {
-        this.isoWeekday(1);
-    }
-
-    // quarters are also special
-    if (units === 'quarter') {
-        this.month(Math.floor(this.month() / 3) * 3);
+                break;
+        }
+    } else {
+        switch (units) {
+            case 'year':
+                this._d.setTime(new Date(this.year(), 0).valueOf());
+                hooks.updateOffset(this, true);
+                break;
+            case 'quarter':
+                this._d.setTime(new Date(this.year(), this.month() - (this.month() % 3)).valueOf());
+                hooks.updateOffset(this, true);
+                break;
+            case 'month':
+                this._d.setTime(new Date(this.year(), this.month()).valueOf());
+                hooks.updateOffset(this, true);
+                break;
+            case 'week':
+            case 'isoWeek':
+            case 'day':
+            case 'date':
+                this._d.setTime(new Date(this.year(), this.month(), this.date()).valueOf());
+                hooks.updateOffset(this, true);
+                // weeks are a special case
+                if (units === 'week') {
+                    this.weekday(0);
+                } else if (units === 'isoWeek') {
+                    this.isoWeekday(1);
+                }
+                break;
+            case 'hour':
+                time = this._d.valueOf();
+                this._d.setTime(time - mod(time + this.utcOffset() * MS_PER_MINUTE, MS_PER_HOUR));
+                hooks.updateOffset(this, true);
+                break;
+            case 'minute':
+                time = this._d.valueOf();
+                this._d.setTime(time - mod(time, MS_PER_MINUTE));
+                break;
+            case 'second':
+                time = this._d.valueOf();
+                this._d.setTime(time - mod(time, MS_PER_SECOND));
+                break;
+        }
     }
 
     return this;
 }
 
 export function endOf (units) {
+    var time;
     units = normalizeUnits(units);
-    if (units === undefined || units === 'millisecond') {
+    if (units === undefined || units === 'millisecond' || !this.isValid()) {
         return this;
     }
 
-    // 'date' is an alias for 'day', so it should be considered as such.
-    if (units === 'date') {
-        units = 'day';
-    }
-
-    if (units === 'day') {
-        if (this.isValid()) {
-            if (this._isUTC) {
-                var time = this._d.valueOf();
-                this._d.setTime(time - mod(time, MS_PER_24_HOUR_DAY) + MS_PER_24_HOUR_DAY - 1);
-            } else {
-                this._d.setTime(new Date(this.year(), this.month(), this.date() + 1).valueOf() - 1);
-            }
+    switch (units) {
+        case 'year':
+            this._d.setTime((this._isUTC ? Date.UTC(this.year() + 1, 0) : new Date(this.year() + 1, 0).valueOf()) - 1);
             hooks.updateOffset(this, true);
-        }
-        return this;
+            break;
+        case 'quarter':
+            this._d.setTime(
+                (this._isUTC ?
+                    Date.UTC(this.year(), 3 + this.month() - this.month() % 3) :
+                    new Date(this.year(), 3 + this.month() - this.month() % 3).valueOf()) - 1
+            );
+            hooks.updateOffset(this, true);
+            break;
+        case 'month':
+            this._d.setTime(
+                (this._isUTC ?
+                    Date.UTC(this.year(), 1 + this.month()) :
+                    new Date(this.year(), 1 + this.month()).valueOf()) - 1
+            );
+            hooks.updateOffset(this, true);
+            break;
+        case 'week':
+            // start of week should be: this.date() - this.weekday(), so end of week would be: this.date() - this.weekday() + 7
+            this._d.setTime(
+                (this._isUTC ?
+                    Date.UTC(this.year(), this.month(), this.date() + 7 - this.weekday()) :
+                    new Date(this.year(), this.month(), this.date() + 7 - this.weekday()).valueOf()) - 1
+            );
+            hooks.updateOffset(this, true);
+            break;
+        case 'isoWeek':
+            // start of iso week should be: this.date() - (this.isoWeekday - 1), so end would be: this.date() - (this.isoWeekday() - 1) + 7
+            this._d.setTime(
+                (this._isUTC ?
+                    Date.UTC(this.year(), this.month(), this.date() + 7 - (this.isoWeekday() - 1)) :
+                    new Date(this.year(), this.month(), this.date() + 7 - (this.isoWeekday() - 1)).valueOf()) - 1
+            );
+            hooks.updateOffset(this, true);
+            break;
+        case 'day':
+        case 'date':
+            time = this._d.valueOf();
+            this._d.setTime(
+                (this._isUTC ?
+                    time - mod(time, MS_PER_24_HOUR_DAY) + MS_PER_24_HOUR_DAY :
+                    new Date(this.year(), this.month(), this.date() + 1).valueOf()) - 1
+            );
+            hooks.updateOffset(this, true);
+            break;
+        case 'hour':
+            time = this._d.valueOf();
+            this._d.setTime(time - mod(time + (this._isUTC ? 0 : this.utcOffset() * MS_PER_MINUTE), MS_PER_HOUR) + MS_PER_HOUR - 1);
+            hooks.updateOffset(this, true);
+            break;
+        case 'minute':
+            time = this._d.valueOf();
+            this._d.setTime(time - mod(time, MS_PER_MINUTE) + MS_PER_MINUTE - 1);
+            break;
+        case 'second':
+            time = this._d.valueOf();
+            this._d.setTime(time - mod(time, MS_PER_SECOND) + MS_PER_SECOND - 1);
+            break;
     }
 
-    return this.startOf(units).add(1, (units === 'isoWeek' ? 'week' : units)).subtract(1, 'ms');
+    return this;
 }