--- a/src/lib/moment/start-end-of.js
+++ b/src/lib/moment/start-end-of.js
@@ -13,7 +13,8 @@ function mod(dividend, divisor) {
 
 function localStartOfDate(y, m, d) {
     // the date constructor remaps years 0-99 to 1900-1999
-    if (0 <= y && y <= 99) {
+    if (y < 100 && y >= 0) {
+        // preserve leap years using a full 400 year cycle, then reset
         return new Date(y + 400, m, d) - MS_PER_400_YEARS;
     } else {
         return new Date(y, m, d).valueOf();
@@ -22,7 +23,8 @@ function localStartOfDate(y, m, d) {
 
 function utcStartOfDate(y, m, d) {
     // Date.UTC remaps years 0-99 to 1900-1999
-    if (0 <= y && y <= 99) {
+    if (y < 100 && y >= 0) {
+        // preserve leap years using a full 400 year cycle, then reset
         return Date.UTC(y + 400, m, d) - MS_PER_400_YEARS;
     } else {
         return Date.UTC(y, m, d);