These changes to `startOf` and `endOf` fix several DST-related issues:
- avoid using setHours/setMinutes/setSeconds - fixes #1990, fixes #2749, fixes #3580
- use `Date.UTC` / `new Date` for units bigger than hours - fixes #3132, fixes #4152

Additionally, the changes supersede these pull requests: #3620, #4164 and #4254 

Benchmark results

| Benchmark | moment 2.19.3 | this PR | improvement |
| --- | --- | --- | --- |
| startOf second | 2,598,171 ops/sec ±1.24% | 5,322,216 ops/sec ±1.56% | 2.0x |
| startOf minute | 1,606,046 ops/sec ±0.98% | 7,397,801 ops/sec ±0.92% | 4.6x |
| startOf hour | 1,069,176 ops/sec ±0.96% | 5,015,984 ops/sec ±0.91%   | 4.6x |
| startOf date | 811,815 ops/sec ±1.08% | 1,162,031 ops/sec ±0.89%     | 1.4x |
| startOf day | 827,911 ops/sec ±0.83% | 1,170,806 ops/sec ±1.25%      | 1.4x |
| startOf isoWeek | 274,302 ops/sec ±1.06% | 973,431 ops/sec ±1.09%    | 3.5x |
| startOf week | 298,240 ops/sec ±1.00% | 1,188,010 ops/sec ±0.94%     | 3.9x |
| startOf month | 653,227 ops/sec ±0.85% | 1,447,224 ops/sec ±0.76%    | 2.2x |
| startOf quarter | 381,496 ops/sec ±0.85% | 1,221,562 ops/sec ±0.93%  | 3.2x |
| startOf year | 425,796 ops/sec ±0.98% | 2,047,534 ops/sec ±1.02%     | 4.8x |
| endOf second | 200,544 ops/sec ±0.92% | 5,496,719 ops/sec ±1.00%     | 27.4x |
| endOf minute | 190,097 ops/sec ±1.02% | 6,600,219 ops/sec ±0.94%     | 34.7x |
| endOf hour | 331,129 ops/sec ±0.86% | 4,612,042 ops/sec ±0.88%       | 13.9x |
| endOf date | 249,577 ops/sec ±1.10% | 1,239,091 ops/sec ±0.96%       | 4.9x  |
| endOf day | 251,496 ops/sec ±1.01% | 1,240,042 ops/sec ±1.35%        | 4.9x  |
| endOf isoWeek | 160,238 ops/sec ±1.73% | 917,050 ops/sec ±0.85%      | 5.7x  |
| endOf week | 174,334 ops/sec ±2.02% | 1,216,138 ops/sec ±0.88%       | 6.9x  |
| endOf month | 200,123 ops/sec ±1.24% | 1,480,412 ops/sec ±2.05%      | 7.3x  |
| endOf quarter | 159,346 ops/sec ±0.99% | 1,165,512 ops/sec ±1.02%    | 7.3x  |
| endOf year | 165,736 ops/sec ±1.70% | 2,026,810 ops/sec ±1.67%       | 12.2x |

Additionally, fixed one more bug: `moment('0000-02-29')` was created in year 1900 then reverted to year 0.  Unfortunately Feb 29th doesn't exist in 1900, so the moment ended up as `0000-03-01` instead.

New tests have been added to `start_end_of.js` that would have failed in earlier releases.