@icambron 2017-12-05
This looks great. Are these benchmarks from Moment's benchmarks suite or separate?