@ashsearle 2017-12-05
From moment's benchmark suite - by running `grunt release` then the relevant benchmarks.