@ashsearle 2017-12-22
@ichernev Got any examples or issue IDs for the DST changes you think won't be handled at start of month / day?