@Havunen 2018-03-22
This PR indeed seems to fix issue DST startOf / endOf. Can this be merged?