@jarredt 2018-11-07
Bumping this PR to see if it can make it into an upcoming release. The DST issue caused non-trivial problems for some of our users in Brazil this week when DST rolled over on Sunday. Difficult for us to work around this in our own code. Would like to avoid getting snagged by this the next time DST rolls over there, if possible...